+++ packages/core/react/use-completion.ts
@@ -146,6 +146,7 @@ export function useCompletion({
       onError,
       setError,
       streamData,
+      streamMode,
       mutateStreamData,
     ],
   );